Revert "💄 Overall file repo styling touchups (#450)"
This reverts commit 12875cd2e83c15e32fbbcdfa49a18ab688a23af8.
