+++ src/components/FileRepo/AggregationSidebar/AggregationSidebar.js
@@ -17,14 +17,10 @@ import InfoIcon from 'icons/InfoIcon';
 import { ActionButton } from 'uikit/Button';
 import { TRACKING_EVENTS } from 'services/analyticsTracking';
 import { FilterInput } from 'uikit/Input';
-import Column from 'uikit/Column';
-import Row from 'uikit/Row';
 
 const AggregationWrapper = styled('div')`
   height: 100%;
-  width: calc(20% + ${({ scrollbarWidth }) => scrollbarWidth}px);
-  max-width: ${({ scrollbarWidth }) => 300 + scrollbarWidth}px;
-  min-width: ${({ scrollbarWidth }) => 200 + scrollbarWidth}px;
+  width: ${({ scrollbarWidth }) => 300 + scrollbarWidth}px;
   overflow-y: auto;
   box-shadow: 0 0 4.9px 0.2px ${({ theme }) => theme.shadow};
   border-color: ${({ theme }) => theme.greyScale5};
@@ -44,12 +40,6 @@ const AggregationTitle = styled('div')`
   flex-grow: 1;
   font-size: 18px;
   color: ${({ theme }) => theme.secondary};
-`;
-
-const IdFilterContainer = styled(Column)`
-  .quick-search {
-    margin-bottom: 10px;
-  }
 `;
 
 const AggregationSidebar = compose(injectState, withTheme)(
@@ -130,7 +120,7 @@ const AggregationSidebar = compose(injectState, withTheme)(
               <Trans css={theme.uppercase}>All Filters</Trans>
             </ActionButton>
           </AggregationHeader>
-          <IdFilterContainer className="aggregation-card">
+          <div className="aggregation-card">
             <QuickSearch
               {...{ ...props, setSQON, translateSQONValue }}
               InputComponent={FilterInput}
@@ -144,10 +134,8 @@ const AggregationSidebar = compose(injectState, withTheme)(
                 />
               }
             />
-            <Row justifyContent="flex-end">
-              <UploadIdsButton {...{ theme, effects, state, setSQON, ...props }} />
-            </Row>
-          </IdFilterContainer>
+            <UploadIdsButton {...{ theme, effects, state, setSQON, ...props }} />
+          </div>
           <Aggregations
             {...{
               ...props,
