+++ src/components/FileRepoSidebar/FileRepoSidebar.js
@@ -17,8 +17,7 @@ const Slideable = styled('div')`
   position: relative;
   transition: all 0.25s;
   width: ${({ expanded, containerWidth, contentSidePadding }) =>
-    expanded ? `calc(${containerWidth} + ${contentSidePadding * 2}px)` : '40px'};
-  max-width: 300px;
+    expanded ? `${containerWidth + contentSidePadding * 2}px` : '40px'};
   overflow: hidden;
   box-shadow: 0 0 4.9px 0.2px ${({ theme }) => theme.shadow};
 `;
@@ -27,7 +26,7 @@ const Container = styled('div')`
   overflow-y: auto;
   flex-grow: 0;
   flex-shrink: 1;
-  width: 100%;
+  width: ${({ containerWidth, contentSidePadding }) => containerWidth + contentSidePadding * 2}px;
   min-width: 265px;
   height: 100%;
   background: ${({ theme }) => theme.backgroundGrey};
@@ -74,7 +73,7 @@ const FileRepoSidebar = compose(withTheme, withState('expanded', 'setExpanded', 
     expanded,
     setExpanded,
     style,
-    containerWidth = '25%',
+    containerWidth = 310,
     contentSidePadding = 15,
     ...props
   }) => (
