+++ src/components/Page.js
@@ -1,33 +1,49 @@
 import React from 'react';
 import Header from 'components/Header';
 import Footer from 'components/Footer';
-import styled from 'react-emotion';
-import Column from 'uikit/Column';
 
-const Container = styled(Column)`
-  position: relative;
-  height: 100vh;
-  min-width: 1024px;
-  width: 100%;
-  background-repeat: repeat;
-`;
-
-const ContentWrapper = styled('div')`
-  height: 100%;
-  overflow-y: auto;
-  & > * {
-    height: 100%;
-  }
-`;
-
-const Page = ({ Head = Header, Foot = Footer, Component, ...props }) => (
-  <Container>
-    <Head />
-    <ContentWrapper>
-      <Component {...props} />
-    </ContentWrapper>
-    <Foot />
-  </Container>
+const Page = ({
+  Head = Header,
+  Foot = Footer,
+  Component,
+  backgroundImageUrl,
+  containerStyle,
+  ...props
+}) => (
+  <div
+    css={`
+      position: relative;
+      height: 100vh;
+      min-width: 1024px;
+      background-image: url(${backgroundImageUrl});
+      ${containerStyle};
+    `}
+  >
+    <div
+      css={`
+        background-repeat: repeat;
+        height: 100%;
+        width: 100%;
+        display: flex;
+        flex-direction: column;
+        background-image: linear-gradient(to bottom, #fff 400px, transparent 100%);
+      `}
+    >
+      <Head />
+      <div
+        css={`
+          height: 100%;
+          overflow-y: auto;
+          & > * {
+            height: 100%;
+          }
+        `}
+      >
+        <Component {...props} />
+      </div>
+      <Foot />
+    </div>
+  </div>
 );
 
 export default Page;
