+++ src/components/ShareSaveQuery/ui.js
@@ -4,10 +4,9 @@ import Row from 'uikit/Row';
 
 export const ButtonContainer = styled(Row)`
   ${({ theme }) => theme.center};
-  padding: 10px 5px;
+  padding: 10px;
 `;
 
 export const CustomLightButotn = styled(LightButton)`
-  min-width: 75px;
-  font-size: 10px;
+  min-width: 65px;
 `;
